Move lock configurable back to bundle object
