+++ bundle/deploy/lock/acquire.go
@@ -32,7 +32,7 @@ func (m *acquire) init(b *bundle.Bundle) error {
 
 func (m *acquire) Apply(ctx context.Context, b *bundle.Bundle) ([]bundle.Mutator, error) {
 	// Return early if locking is disabled.
-	if !b.Config.Workspace.Lock.IsEnabled() {
+	if !b.Config.Bundle.Lock.IsEnabled() {
 		log.Infof(ctx, "Skipping; locking is disabled")
 		return nil, nil
 	}
@@ -42,7 +42,7 @@ func (m *acquire) Apply(ctx context.Context, b *bundle.Bundle) ([]bundle.Mutator
 		return nil, err
 	}
 
-	force := b.Config.Workspace.Lock.Force
+	force := b.Config.Bundle.Lock.Force
 	log.Infof(ctx, "Acquiring deployment lock (force: %v)", force)
 	err = b.Locker.Lock(ctx, force)
 	if err != nil {
